system/extras: Fix for maps parsing bug
Fix for a bug where a line in the maps file that is greater than 256
characters was causing invalid entries to be added to the maps list.
The size of the line buffer has been increased to accomodate longer
maps entries.

Basically what was happening was that only the first 256 characters
of the maps line were being considered as the maps information, and
on the next iteration of the while loop the remaining characters in
the line were being considered as a separate entry in the maps file,
which is obviously not correct.

Change-Id: I35a06d5d1c82c62a807f93cb8e7d68b352f54f68